Add in all missing Index methods
